backend/pyproject.toml
[project]
name = "cyclecast-backend"
version = "0.1.0"
description = "CycleCast Backend API Proxy"
authors = [
    { name = "Dan Higgins" }
]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
